station/config.py
# station/config.py
import os
import yaml

def load_config(config_path="config.yml"):
    """Load configuration from a YAML file."""
    try:
        with open(config_path, "r") as file:
            return yaml.safe_load(file)
    except FileNotFoundError:
        print(f"Configuration file not found: {config_path}")
        return {}
